[#83] - Use coinUiModel instead of coinDetail - Move HomeScreenUITest to home folder - Rename DetailScreenUiTest to DetailScreenUITest
